gpt-telegram-bot/requirements.txt
python-telegram-bot>=21.0
google-generativeai>=0.8.0
python-dotenv>=1.0